Added the ability to append new frames to an animation. Also added Onion Skins!
-The Animation Timeline now has a button (with a temporary layout) that will add a new frame to the animation

-Reorganized a few things to make use of the specialized view and view model classes for the Animation Timeline that I originally added last night but didn't have time to hook up

-Added two additional InkCanvas controls to the WorkspaceView. Eventually this will probably need to be carved off into a more specialized series of Views and ViewModels just for the drawing canvas area of the workspace view. I also added a pair of additional StrokeCollections to the AnimationTimelineViewModel. One for the onion skinned collection of strokes from the frame immediately preceeding the selected frame, and another similar collection for the strokes from the frame immediately following the selected frame.
These are bound to the two new InkCanvas controls and everything is set up so that the strokes are drawn in either red or green half opaque strokes depending on whether they are strokes from the previous or next frame. Ultimately, I'd like to change the two InkCanvas instances to an ItemTemplate used on any items in the bound collection of an Canvas or other ItemsControl. Each item in that collection would represent the strokes from a rolling set of frames which come immediately before or after the selected frame.
As it stands, all of the strokes from any previous or next frames must be copied into a new collection so that the drawing attributes of each stroke can be modified so that they are semi transparent green or red, thus signalling that they are not strokes from the currently selected frame and are therefore not editable.
There is already a property defined on the WorkspaceViewModel that can keep track of OnionSkin visibility. Right now, it's an enum that can track not just whether the onion skin is turned on or not, but also whether it will only apply to frames that come after the current frame, before the current frame, or both. But this should A) eventually be moved into a specialized class for ViewSettingStates, B) be used in concert with a few other settings for the number of frames that will be onion skinned, and possibly C) be split, and simplified so that a single boolean stores the on/off state of the OnionSkin itself, and a separate enum stores the type of onion skins (previous,next, both) and then another property stores how many frames in any of the indicated directions will be onion skinned. It also may be worthwhile to eventually allow the color and starting opacity of onionskin strokes to be editable by the user.