Added Selection to the EditorTool Panel
-Since we're using the WPF InkCanvas control, there are some things that are baked into the Selection tool that are worth mentioning.
First, the selection tool is weird. It's functions like a sort of quasi-lasso selection. Meaning it seems to behave largely like a Lasso Select, but it doesn't require a closed area in order to complete a selection. From what I can tell, it seems almost like it automatically, and invisibly closes the selection area the moment a third point is placed along the selection path. To clarify, given the ordered points A, B, and C of a selection path, it will, internally, treat points A and C as though they are connected, creating a complete perimeter around an area that it treats as a selection target, and any strokes for which all constituent points reside completely within that target area will be selected. The target area is constantly updated by repeatedly connecting the first and last points in the selection path any time a new point is appended to the existing path. The end result of all of this is that it feels like it's constantly lighting things up with the selection visual cue before you expect it to.
-It doesn't seem to care much for a selection path that results in degenerate geometry. Again, this explanation relies on a few basic assumptions that all seem reasonable enough given the graphics formats that show up most frequently in WPF and based on the way the tool appears to behave in certain test cases. With that in mind, it isn't surprising that, when given a selection path that crosses over itself in one or more places, it generally doesn't result in the selection of strokes just because they happen to fall entirely within a rectangle created from the minimum and maximum points of the path. What's probably happening here is that when the path crosses over itself, it "flips" the shape over in that spot, and since the geometry is only "1-sided" anyplace that the reverse side would face the user, is simply treated like it's invisible. This means that if you create a single path that takes on a roughly |X| shape, only one of the resulting triangular regions will select a stroke within it's borders. This is made even less predictable by the above mentioned behavior whereby the first and last points are treated as connected for the purpose of calculating the selection area.
-Resize and Move Selection Features come free with Selection, and they behave largely as you'd expect.