MainWIndow WorkspaceManager is a static instance now.
NOTE: This might have a fairly problematic side effect down the line. It seems like making this change means that some controls on the initial blank WorkspaceView throw an exception if they use a style that is based on another style that is also defined as a StaticResource. I think it might have something to do with the order in which the styles are loaded, but I'm not positive about this. At any rate it should be mentioned just in case it becomes a problem later, we can change how the WorkspaceManager is instantiated if this causes issues down the line.

-I also did some staging for the upcoming Undo/Redo changes.
-Added Undo/Redo buttons to the toolbar
-Added blank commands for each
-Added a few of the interfaces for the Memento pattern that the Undo system will use
-Hooked up all of the above mentioned stubs