Add stylus points changed handler to every stroke
This fixes the bug that moving a stroke did not create an undo record.
That bug was first fixed in commit 307d974c7a6088a8ba332920eb53770cc95a9ecf
but this shows the precise code which fixes the bug.

(Probably several changes work together to *fully* fix this bug, as this
is probably the issue which started the work that went into 307d974.)