Bugfix: Attempting to Change the Brush Color crashes the application
-So, this one is interesting, and I'm not 100% sure that what appears to have happened is what actually happened, but I'm going to do my best to explain what my current understanding of it is.

-It appears that, as of .NET Core 3.1, that the Working Directory of an application that's run from Visual Studio no longer defaults to ProjectDirectory/bin/Debug (or Release as the case may be). Presumably, this is also related to the fact that, in fact, the assembly isn't deployed specifically to that directory at all. Instead, the build output is sent to /bin/Debug/netcoreapp3.1
Although I haven't actually tested it, I think it's reasonable to assume that the last directory in that path changes depending on the version of .Net Core targeted by the build. In any case, since the default build output directory has changed, it was a bit less surprising to learn that the project default working directory was also changed, but it was a somewhat surprising to learn that it is no longer set to the same directory as build output. That being the case though, even relative paths that worked fine in the original .NET version of this project began throwing some manner of file not found exception when the project was switched to .NET Core. In this case, those exceptions were in reference to the [ProjectDirectory]/Resources/Cursors/ColorPickerCursor.cur file.

-Ultimately, in this case, the fix was relatively simple. I simply had to change the file properties in the Solution Explorer so that Build Action was set to Resource, and Copy To Output Directory was set to Always. I wish I had any real understanding of how this materially changes things, but all of the above is really nothing more than a guess, so if it's not actually what's going on, then I don't have any idea what was causing this, or why it works now.